Refactor notifications redux store section
