+++ src/selectors/index.ts
@@ -61,9 +61,6 @@ export const getMyAccounts: (state: RootState) => ReadonlyArray<ChainAccount> = 
     })),
 );
 
-export const getPendingTransactions = (state: RootState) => state.notification.pending || [];
-export const getTransactions = (state: RootState) => state.notification.transaction || [];
-
 /* TODO add some generic "require" helper? */
 export const requireActiveIdentity = (state: RootState) => {
   const ident = getActiveIdentity(state);
