+++ src/store/notifications/selectors.ts
@@ -1,0 +1,4 @@
+import { RootState } from "~/reducers";
+
+export const getPendingTransactions = (state: RootState) => state.notification.pending || [];
+export const getTransactions = (state: RootState) => state.notification.transaction || [];
